+++ scripts/git-subtree/jupyter-client/add.ps1
@@ -1,5 +1,5 @@
 git subtree add `
     -P workspace/plugins/jupyter-client `
     https://github.com/Zuoqiu-Yingyi/siyuan-plugin-jupyter-client.git `
-    main `
+    dev `
     --squash