+++ scripts/git-subtree/jupyter-client/pull.ps1
@@ -1,4 +1,4 @@
 git subtree pull `
     -P workspace/plugins/jupyter-client `
     https://github.com/Zuoqiu-Yingyi/siyuan-plugin-jupyter-client.git `
-    main
+    dev