+++ scripts/git-subtree/jupyter-client/push.ps1
@@ -1,4 +1,5 @@
 git subtree push `
     -P workspace/plugins/jupyter-client `
     https://github.com/Zuoqiu-Yingyi/siyuan-plugin-jupyter-client.git `
-    main
+    dev `
+    -S