+++ scripts/git-subtree/open-api/add.ps1
@@ -1,5 +1,5 @@
 git subtree add `
     -P workspace/plugins/open-api `
     https://github.com/Zuoqiu-Yingyi/siyuan-plugin-open-api.git `
-    main `
+    dev `
     --squash