+++ scripts/git-subtree/open-api/pull.ps1
@@ -1,4 +1,4 @@
 git subtree pull `
     -P workspace/plugins/open-api `
     https://github.com/Zuoqiu-Yingyi/siyuan-plugin-open-api.git `
-    main
+    dev