+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - 调整更改日志格式 | Adjust the format of the change log.
 - 添加 `release-please.yml` CD 配置 | Add `release-please.yml` CD configuration.
 - 更新依赖 | Update dependencies.
+- 更新 subtree 管理脚本 | Update subtree management scripts.
 
 ## 2023-05-18